requirements.txt
google-adk==1.14.0

fastapi
uvicorn
pydantic
python-dotenv

google-cloud-logging
google-cloud-datastore
google-cloud-bigquery

google-auth
google-auth-oauthlib
google-auth-httplib2

langchain
langchain-core
langchain-community==0.3.27
wikipedia==1.4.0

pandas
numpy